fix props (pos/size/...) in render for non-integer zoom, fix non-integer spacing for vbox/hbox, fix warnings

+++ src/gui/screen/container.h
@@ -8,7 +8,7 @@ public:
 	bool hasVBox = false;
 	bool hasHBox = false;
 
-	int spacing = 0;
+	float spacing = 0;
 
 	std::vector<Child*> screenChildren;
